+++ MyProject/MyProject.Data/Models/Setting.cs
@@ -35,6 +35,6 @@ namespace MyProject.Data.Models
 
         public bool? IsDeleted { get; set; } = false;
 
-        public int? BusinessId { get; set; }
+        public long? BusinessId { get; set; }
     }
 }
